Worked on Timer again, problems with prefab (see description)
Timer and FinishedTime is now the same in each level. There is still a problem with the GameplayManager in the prefabs, because I cannot add the timer, since it is inside another object. Don't know how to solve it, yet. I added manually the timer to level 1-5 it works fine for those.