Loxer dev (#6)
* Timer, MainMenu, ScreenMenu

* minor fixes

added a boolean in PlayerController if the player can move or not (turns to false when a level has been finished)

* Formatting

* Merge Scenes and Prefabs

* Fixed respawn-movement-problem

* All levels playable from start + Movement adjustments

* Rope anim + fixes

Fixed mouse position/ crosshair
Fixed moving objects take the anchor with them

* Added shake + better goal

* Timer restarts after respawn again

* Finished time is visible again

* Worked on Timer again, problems with prefab (see description)

Timer and FinishedTime is now the same in each level. There is still a problem with the GameplayManager in the prefabs, because I cannot add the timer, since it is inside another object. Don't know how to solve it, yet. I added manually the timer to level 1-5 it works fine for those.

* Changed some logic of Timer as preparation for saving the time

* Update Timer.cs

* Shows if there is a new BestTime, although it's not saving it, yet

* Timer works now in all levels

* Formatting

* Update ScoreMenu.cs

* Update Timer.cs

* Local Save Games are finally working

* Update GameplayManager.cs

* Update ScoreMenu.cs

* Update Timer.cs